lab-12/RPi-YOLOv10/RPi-YOLOv10-main/requirements.txt
numpy
ai-edge-litert
opencv-python-headless
paho-mqtt==1.6.1